requirements.txt
streamlit
pandas
numpy
scipy
statsmodels
plotly
python-dotenv
groq
openpyxl